# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmAid API
  description: Disaster relief coordination platform connecting surplus resources with needs
  version: 1.0.0
  contact:
    name: SwarmAid Team
    email: api@swarmaid.org
servers:
  - url: https://api.swarmaid.org
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Coordinates:
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
      required: [lat, lon]
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required: [street, city, state, postalCode, country]
    TimeWindow:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required: [start, end]
...
    DemandPost:
      type: object
      properties:
        id:
          type: string
        recipientId:
          type: string
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        quantityNeeded:
          type: number
        quantityUnit:
          type: string
        capacity:
          type: number
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        deliveryCoordinates:
          $ref: '#/components/schemas/Coordinates'
        acceptanceWindow:
          $ref: '#/components/schemas/TimeWindow'
        specialRequirements:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [posted, matched, scheduled, picked_up, delivered, closed, expired]
    MatchRecommendation:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        demandId:
          type: string
        supplierId:
          type: string
        recipientId:
          type: string
        score:
          type: number
        scoreBreakdown:
          type: object
        distanceMiles:
          type: number
        estimatedDurationMinutes:
          type: number
        status:
          type: string
          enum: [pending, accepted, rejected, scheduled, in_progress, completed, failed]
        routePlan:
          type: object
        complianceStatus:
          type: string
          enum: [pending, passed, blocked, override]
...
paths:
...
  /demand:
    post:
      summary: Create demand post
      tags: [Demand]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
                quantityNeeded:
                  type: number
                quantityUnit:
                  type: string
                capacity:
                  type: number
                deliveryAddress:
                  $ref: '#/components/schemas/Address'
                acceptanceWindow:
                  $ref: '#/components/schemas/TimeWindow'
                specialRequirements:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Demand post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandPost'

    get:
      summary: List demand posts
      tags: [Demand]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of demand posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  demands:
                    type: array
                    items:
                      $ref: '#/components/schemas/DemandPost'
...
  /matches/recommendations:
    post:
      summary: Create match recommendations
      tags: [Matches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listingId:
                  type: string
                demandId:
                  type: string
      responses:
        '201':
          description: Recommendations created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRecommendation'
  /matches:
    get:
      summary: List match recommendations
      tags: [Matches]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRecommendation'
...